Fixes #10 by putting UserDetailsService in a separate class
