+++ StockWatch.Backend/src/main/java/com/henning/oden/java/StockAlert/Backend/services/CustomUserDetailsService.java
@@ -1,16 +1,47 @@
 package com.henning.oden.java.StockAlert.Backend.services;
 
+import com.henning.oden.java.StockAlert.Backend.entities.SystemUser;
 import com.henning.oden.java.StockAlert.Backend.repos.SystemUserRepository;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.provisioning.InMemoryUserDetailsManager;
+import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
+import java.util.Collection;
+
+@Service
 public class CustomUserDetailsService implements UserDetailsService {
 
     private SystemUserRepository users;
 
-    public CustomUserDetailsService(SystemUserRepository users) {
+    private PasswordEncoder passwordEncoder;
+    public CustomUserDetailsService(SystemUserRepository users, PasswordEncoder passwordEncoder) {
         this.users = users;
+        this.passwordEncoder = passwordEncoder;
+    }
+
+    @Bean
+    public UserDetailsService userDetailsService() {
+
+        GrantedAuthority testAuth = new SimpleGrantedAuthority("ROLE_USER");
+
+        Collection<GrantedAuthority> authorities = new ArrayList<>();
+
+        authorities.add(testAuth);
+
+        SystemUser testSystemUser = new SystemUser("Test", passwordEncoder.encode("password"), authorities);
+
+        InMemoryUserDetailsManager userDetailsManager = new InMemoryUserDetailsManager();
+
+        userDetailsManager.createUser(testSystemUser);
+
+        return userDetailsManager;
     }
 
     @Override
